feat: 이메일 인증번호 확인 구현

+++ src/main/java/kwthon_1team/kwthon/service/MemberService.java
@@ -1,6 +1,6 @@
 package kwthon_1team.kwthon.service;
 
-import kwthon_1team.kwthon.common.BaseErrorResponse;
+import jakarta.servlet.http.HttpSession;
 import kwthon_1team.kwthon.domian.dto.request.AuthRequestDto;
 import kwthon_1team.kwthon.domian.entity.EmailVerification;
 import kwthon_1team.kwthon.domian.entity.Member;
@@ -11,9 +11,10 @@ import kwthon_1team.kwthon.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
 import java.time.LocalDateTime;
-import java.util.Random;
 
 @Service
 @Transactional
@@ -27,22 +28,22 @@ public class MemberService {
     public void join(AuthRequestDto authRequestDto) {
         validateAuthRequest(authRequestDto);
         checkDuplicateEmail(authRequestDto.getEmail());
-        // verifyEmailCode(authRequestDto.getEmail(), Integer.valueOf(authRequestDto.getVerificationCode()));
         memberRepository.save(converToMember(authRequestDto));
     }
 
-    @Transactional (readOnly = true)
+    @Transactional
     public void checkDuplicateEmail(String email) {
         String trimmedEmail = email.trim();
         if (!memberRepository.findByEmail(trimmedEmail).stream().toList().isEmpty()) {
             throw new BaseException(400, "이미 회원가입된 이메일");
         }
         validateUniversityEmail(trimmedEmail);
+
+        requestEmailVerification(trimmedEmail);
     }
 
     @Transactional (readOnly = true)
     public void validateUniversityEmail(String email) {
-        System.out.println("이메일 도메인 체크하겠음");
         if (!email.endsWith("@kw.ac.kr")) {
             throw new BaseException(400, "광운대학교 이메일(@kw.ac.kr)만 입력해 주세요.");
         }
@@ -50,17 +51,17 @@ public class MemberService {
 
     private void validateAuthRequest(AuthRequestDto authRequestDto) {
         if (authRequestDto.getStudentId() == null)
-            throw new BadRequestException("아이디를 입력해 주세요.");
+            throw new BaseException(400, "아이디를 입력해 주세요.");
         if (authRequestDto.getDepartment() == null)
-            throw new BadRequestException("학과를 입력해 주세요.");
+            throw new BaseException(400, "학과를 입력해 주세요.");
         if (authRequestDto.getPassword() == null)
-            throw new BadRequestException("비밀번호를 입력해 주세요.");
+            throw new BaseException(400, "비밀번호를 입력해 주세요.");
         if (authRequestDto.getName() == null)
-            throw new BadRequestException("이름을 입력해 주세요.");
+            throw new BaseException(400, "이름을 입력해 주세요.");
         if (authRequestDto.getEmail() == null)
-            throw new BadRequestException("이메일을 입력해 주세요.");
+            throw new BaseException(400, "이메일을 입력해 주세요.");
         if (!authRequestDto.isAgreement())
-            throw new BadRequestException("개인정보 수집에 동의해 주세요.");
+            throw new BaseException(400, "개인정보 수집에 동의해 주세요.");
     }
 
     private Member converToMember(AuthRequestDto authRequestDto) {
@@ -69,9 +70,6 @@ public class MemberService {
 
     @Transactional
     public void requestEmailVerification(String email) {
-        validateUniversityEmail(email);
-        checkDuplicateEmail(email);
-
         // 인증 코드 생성 및 저장
         Integer verificationCode = generateVerificationCode();
         LocalDateTime expirationDate = LocalDateTime.now().plusMinutes(10);  // 유효 기간 10분
@@ -79,6 +77,8 @@ public class MemberService {
         emailVerificationRepository.save(emailVerification);
 
         emailService.sendValidateEmailRequestMessage(email, verificationCode.toString());
+
+        saveEmailToSession(email);
     }
 
     private Integer generateVerificationCode() {
@@ -86,14 +86,40 @@ public class MemberService {
     }
 
     @Transactional
-    public void verifyEmailCode(String email, Integer code) {
+    public void verifyEmailCode(Integer code) {
+        String email = getEmailFromSession();
+
         EmailVerification emailVerification = emailVerificationRepository.findByEmail(email)
-                .orElseThrow(()->new BadRequestException("이메일 인증이 필요합니다."));
+                .orElseThrow(()->new BaseException(400, "이메일 인증이 필요합니다."));
+
+        // 인증 코드 유효 기간 검증
+        if (emailVerification.getExpirationTime().isBefore(LocalDateTime.now())) {
+            throw new BaseException(400, "인증번호가 만료되었습니다.");
+        }
 
         if (emailVerification.getVerificationCode() == null ||
             !emailVerification.getVerificationCode().equals(code)) {
-            throw new BadRequestException("인증번호가 일치하지 않습니다.");
+            throw new BaseException(400, "인증번호가 일치하지 않습니다.");
         }
         emailVerificationRepository.delete(emailVerification);
+    }
+
+    private void saveEmailToSession(String email) {
+        HttpSession session = getHttpSession();
+        session.setAttribute("email", email);
+    }
+
+    private HttpSession getHttpSession() {
+        ServletRequestAttributes attr = (ServletRequestAttributes) RequestContextHolder.currentRequestAttributes();
+        return attr.getRequest().getSession(true);
+    }
+
+    private String getEmailFromSession() {
+        HttpSession session = getHttpSession();
+        String email = (String) session.getAttribute("email");
+        if (email == null) {
+            throw new BaseException(400, "이메일 세션이 존재하지 않습니다.");
+        }
+        return email;
     }
 }
